Задача 43.

+++ zd43/Program.cs
@@ -1,10 +1,9 @@
 ﻿
-
 //Задача 43: Напишите программу, которая найдёт точку пересечения двух прямых, заданных уравнениями y = k1 * x + b1, y = k2 * x + b2; значения b1, k1, b2 и k2 задаются пользователем.
 
 Console.WriteLine("Задача 41: Пользователь вводит с клавиатуры M чисел. Посчитайте, сколько чисел больше 0 ввёл пользователь.");
 Console.Write("Введите k1:");
-int k1=Convert.ToFloat(Console.ReadLine());
+int k1=Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите b1:");
 int b1=Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите k2:");
